Version bump

+++ server/server.cpp
@@ -52,7 +52,7 @@ void Server::onNewSession(const std::shared_ptr<StreamSession>& session)
 void Server::onMetaChanged(const PcmStream* pcmStream)
 {
     // clang-format off
-    // Notification: {"jsonrpc":"2.0","method":"Stream.OnMetadata","params":{"id":"stream 1", "meta": {"album": "some album", "artist": "some artist", "track": "some track"...}}
+    // Notification: {"jsonrpc":"2.0","method":"Stream.OnMetadata","params":{"id":"stream 1", "meta": {"album": "some album", "artist": "some artist", "track": "some track"...}}}
     // clang-format on
 
     const auto meta = pcmStream->getMeta();
@@ -73,7 +73,7 @@ void Server::onStateChanged(const PcmStream* pcmStream, ReaderState state)
     // clang-format off
     // Notification: {"jsonrpc":"2.0","method":"Stream.OnUpdate","params":{"id":"stream 1","stream":{"id":"stream 1","status":"idle","uri":{"fragment":"","host":"","path":"/tmp/snapfifo","query":{"chunk_ms":"20","codec":"flac","name":"stream 1","sampleformat":"48000:16:2"},"raw":"pipe:///tmp/snapfifo?name=stream 1","scheme":"pipe"}}}}
     // clang-format on
-    LOG(INFO, LOG_TAG) << "onStateChanged (" << pcmStream->getName() << "): " << static_cast<int>(state) << "\n";
+    LOG(INFO, LOG_TAG) << "onStateChanged (" << pcmStream->getName() << "): " << state << "\n";
     //	LOG(INFO, LOG_TAG) << pcmStream->toJson().dump(4);
     json notification = jsonrpcpp::Notification("Stream.OnUpdate", jsonrpcpp::Parameter("id", pcmStream->getId(), "stream", pcmStream->toJson())).to_json();
     controlServer_->send(notification.dump(), nullptr);
@@ -411,7 +411,6 @@ void Server::processRequest(const jsonrpcpp::request_ptr request, jsonrpcpp::ent
                 // clang-format off
                 // Request:      {"id":4,"jsonrpc":"2.0","method":"Stream.SetMeta","params":{"id":"Spotify", "meta": {"album": "some album", "artist": "some artist", "track": "some track"...}}}
                 // Response:     {"id":4,"jsonrpc":"2.0","result":{"stream_id":"Spotify"}}
-                // Call onMetaChanged(const PcmStream* pcmStream) for updates and notifications
                 // clang-format on
 
                 LOG(INFO, LOG_TAG) << "Stream.SetMeta(" << request->params().get<std::string>("id") << ")" << request->params().get("meta") << "\n";
@@ -433,7 +432,6 @@ void Server::processRequest(const jsonrpcpp::request_ptr request, jsonrpcpp::ent
                 // clang-format off
                 // Request:      {"id":4,"jsonrpc":"2.0","method":"Stream.AddStream","params":{"streamUri":"uri"}}
                 // Response:     {"id":4,"jsonrpc":"2.0","result":{"stream_id":"Spotify"}}
-                // Call onMetaChanged(const PcmStream* pcmStream) for updates and notifications
                 // clang-format on
 
                 LOG(INFO, LOG_TAG) << "Stream.AddStream(" << request->params().get("streamUri") << ")"
@@ -453,7 +451,6 @@ void Server::processRequest(const jsonrpcpp::request_ptr request, jsonrpcpp::ent
                 // clang-format off
                 // Request:      {"id":4,"jsonrpc":"2.0","method":"Stream.RemoveStream","params":{"id":"Spotify"}}
                 // Response:     {"id":4,"jsonrpc":"2.0","result":{"stream_id":"Spotify"}}
-                // Call onMetaChanged(const PcmStream* pcmStream) for updates and notifications
                 // clang-format on
 
                 LOG(INFO, LOG_TAG) << "Stream.RemoveStream(" << request->params().get("id") << ")"
